requirements.txt
pandas
python-dotenv
requests
MetaTrader5
python-dateutil
